Persist quality models.
