+++ analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/MetricAggregationOperator.java
@@ -1,258 +1,35 @@
-/**
- */
 package eubrazil.atmosphere.qualitymodel;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * <!-- begin-user-doc -->
- * A representation of the literals of the enumeration '<em><b>Metric Aggregation Operator</b></em>',
- * and utility methods for working with them.
- * <!-- end-user-doc -->
- * @see eubrazil.atmosphere.qualitymodel.QualityModelPackage#getMetricAggregationOperator()
- * @generated
- */
 public enum MetricAggregationOperator {
-	/**
-	 * The '<em><b>AVERAGE</b></em>' literal object.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #AVERAGE_VALUE
-	 * @generated
-	 * @ordered
-	 */
-	AVERAGE(0, "AVERAGE", "AVERAGE"),
+	
+	AVERAGE(0, "AVERAGE"), MINIMUM(1, "MINIMUM"), MAXIMUM(2, "MAXIMUM"), SUM(3, "SUM");
 
-	/**
-	 * The '<em><b>MINIMUM</b></em>' literal object.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #MINIMUM_VALUE
-	 * @generated
-	 * @ordered
-	 */
-	MINIMUM(0, "MINIMUM", "MINIMUM"),
+	private Integer valor;
+	private String label;
 
-	/**
-	 * The '<em><b>MAXIMUM</b></em>' literal object.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #MAXIMUM_VALUE
-	 * @generated
-	 * @ordered
-	 */
-	MAXIMUM(0, "MAXIMUM", "MAXIMUM"),
+	private MetricAggregationOperator(Integer valor, String label) {
+		this.valor = valor;
+		this.label = label;
+	}
 
-	/**
-	 * The '<em><b>SUM</b></em>' literal object.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #SUM_VALUE
-	 * @generated
-	 * @ordered
-	 */
-	SUM(0, "SUM", "SUM");
+	public Integer getValor() {
+		return valor;
+	}
 
-	/**
-	 * The '<em><b>AVERAGE</b></em>' literal value.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of '<em><b>AVERAGE</b></em>' literal object isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @see #AVERAGE
-	 * @generated
-	 * @ordered
-	 */
-	public static final int AVERAGE_VALUE = 0;
+	public String getLabel() {
+		return label;
+	}
 
-	/**
-	 * The '<em><b>MINIMUM</b></em>' literal value.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of '<em><b>MINIMUM</b></em>' literal object isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @see #MINIMUM
-	 * @generated
-	 * @ordered
-	 */
-	public static final int MINIMUM_VALUE = 0;
-
-	/**
-	 * The '<em><b>MAXIMUM</b></em>' literal value.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of '<em><b>MAXIMUM</b></em>' literal object isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @see #MAXIMUM
-	 * @generated
-	 * @ordered
-	 */
-	public static final int MAXIMUM_VALUE = 0;
-
-	/**
-	 * The '<em><b>SUM</b></em>' literal value.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of '<em><b>SUM</b></em>' literal object isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @see #SUM
-	 * @generated
-	 * @ordered
-	 */
-	public static final int SUM_VALUE = 0;
-
-	/**
-	 * An array of all the '<em><b>Metric Aggregation Operator</b></em>' enumerators.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	private static final MetricAggregationOperator[] VALUES_ARRAY =
-		new MetricAggregationOperator[] {
-			AVERAGE,
-			MINIMUM,
-			MAXIMUM,
-			SUM,
-		};
-
-	/**
-	 * A public read-only list of all the '<em><b>Metric Aggregation Operator</b></em>' enumerators.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public static final List<MetricAggregationOperator> VALUES = Collections.unmodifiableList(Arrays.asList(VALUES_ARRAY));
-
-	/**
-	 * Returns the '<em><b>Metric Aggregation Operator</b></em>' literal with the specified literal value.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param literal the literal.
-	 * @return the matching enumerator or <code>null</code>.
-	 * @generated
-	 */
-	public static MetricAggregationOperator get(String literal) {
-		for (int i = 0; i < VALUES_ARRAY.length; ++i) {
-			MetricAggregationOperator result = VALUES_ARRAY[i];
-			if (result.toString().equals(literal)) {
-				return result;
+	public static MetricAggregationOperator getEnumByValor(Integer valor) {
+		if (valor == null) {
+			return null;
+		}
+		for (MetricAggregationOperator item : MetricAggregationOperator.values()) {
+			if (item.getValor().intValue() == valor.intValue()) {
+				return item;
 			}
 		}
 		return null;
-	}
-
-	/**
-	 * Returns the '<em><b>Metric Aggregation Operator</b></em>' literal with the specified name.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param name the name.
-	 * @return the matching enumerator or <code>null</code>.
-	 * @generated
-	 */
-	public static MetricAggregationOperator getByName(String name) {
-		for (int i = 0; i < VALUES_ARRAY.length; ++i) {
-			MetricAggregationOperator result = VALUES_ARRAY[i];
-			if (result.getName().equals(name)) {
-				return result;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Returns the '<em><b>Metric Aggregation Operator</b></em>' literal with the specified integer value.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param value the integer value.
-	 * @return the matching enumerator or <code>null</code>.
-	 * @generated
-	 */
-	public static MetricAggregationOperator get(int value) {
-		switch (value) {
-			case AVERAGE_VALUE: return AVERAGE;
-		}
-		return null;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	private final int value;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	private final String name;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	private final String literal;
-
-	/**
-	 * Only this class can construct instances.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	private MetricAggregationOperator(int value, String name, String literal) {
-		this.value = value;
-		this.name = name;
-		this.literal = literal;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public int getValue() {
-	  return value;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String getName() {
-	  return name;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String getLiteral() {
-	  return literal;
-	}
-
-	/**
-	 * Returns the literal value of the enumerator, which is its string representation.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		return literal;
 	}
 	
-} //MetricAggregationOperator
+}
